Maintain backwards compat w/conduit versions <0.2.1
ReliableChan now uses a protocol to get a com.rabbitmq.client.Connection
object. This enables the client code to pass in anything that implements
com.farmlogs.conduit.RMQConnection. The immediate benefit of this is
that code written to conform to versions <0.2.1 of conduit will still
work with the new reliable chan implemented in 0.2.1.